Change loglevel to WARN when requestHandler throws an exception.
This is when the application's requestHandler throws some kind of
exception when trying to apply a write. Because the exception is
caught and then stuffed in a SiriusResult.error, for the hosting
application to handle, logging at WARN seems appropriate (as the
application might well handle it gracefully). However, since it
does cross a library boundary without being an actual exception,
it is easy to accidentally ignore, so removing the log statement
completely seems like a recipe for hiding problems.

Change-Id: I72590def8136cd3247b759b0e25f188fa9803061
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/4203
Tested-by: GerritAHP
Reviewed-by: kristofer tomasette <kristofer_tomasette@comcast.com>