Introduce SegmentedUberStore with Compaction
Introducing a new SiriusLog implementation, building from the base
UberStore. A SegmentedUberStore uses the same UberStore backend files,
but structures them in directories known as Segments. It splits off
new Segments after a configurable number of events. A newly-split
segment has a numerical "name", which is the previous segment's name +
1.

These segments are divided into two parts:
- inactive, which consist of the read-only files from closed segments
- active, which is where new writes are applied

Compaction consists of:
- for each inactive segment S, where S.keys-applied == false:
  - gather the keys(S)
  - for each inactive segment S', where S'.name < S.name:
    - build S'.compacting, which consists of all events in S' whose keys
      were not found in keys(S)
  - for each inactive segment S', where there exists S'.compacting
    - replace S' with S'.compacting
  - mark S.keys-applied = true

There is no internal compaction of a segment. This may be a future
enhancement.

Change-Id: I7139189a29e20e6949cb049b220e85e0dd551e72
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/4913
Tested-by: GerritAHP
Reviewed-by: byron maulan <maulan_byron@cable.comcast.com>