Use is-legacy to determine whether to compact
compact_wal will still have the side-effect of creating a backup, but it
will no longer compact if the WAL is a Segmented log.

This also changes the way the path to waltool is determined, basing it
on the location of the called script (compact_wal.sh) instead of the
uberstore. I think this makes more sense but am certainly open to
discussion.

Change-Id: I87b67f44d23e3db3f839612f8982b3031afb12cb
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/7729
Tested-by: GerritAHP
Reviewed-by: kristofer tomasette <kristofer_tomasette@comcast.com>
Reviewed-by: Joe <joseph_campbell@comcast.com>