Introduce merge() to SegmentedUberStore
When two adjacent files have an aggregate size less than the
maxEventsPerSegment, and both have had their keys applied in compaction,
they can be merged into one. This introduces that functionality.

After a round of compaction has finished, a round of merges will be
triggered. The time to do this will be included in the "compaction time"
metric. It shouldn't be much in the general case.

Change-Id: I2df569dec86399602a01713fe10f5c89ed9b2ac6
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/7826
Tested-by: GerritAHP
Reviewed-by: kristofer tomasette <kristofer_tomasette@comcast.com>