Convert apply() to props()
In preparation for upgrade. This is the new standard practice for
actors, pushing the Props creation into the actor's companion object.

Also here:
- make Configuration explicit in some places, as a by-product of
  changing the instantiation of some of these guys. There will be
  more on this later.
- finishing up a little bit of ChildProvider refactoring that was
  previously missed.

Sorry there are multiple things in here, but it's mostly wiring
and construction -- so it's fairly straightforward, but the changes
end up being a bit intertwined. I can probably split it further if
necessary, but we'll see if this works.

patchset 2: commented SiriusSupervisor.

Change-Id: Ie90cf839aaa64fa1bb0c42feae0d073f37d62263
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/8431
Tested-by: GerritAHP
Reviewed-by: kristofer tomasette <kristofer_tomasette@comcast.com>