Update Membership to use Map(path => ActorRef)
Just used a Set of ActorRefs before, but this proves to be too unwieldy
when we need to resolve paths into ActorRefs upon instantiation (instead
of just assuming they'll work). This updates the membership
datastructure and algorithm a little bit, and lays a little bit of
ground work for the upcoming akka upgrade.

Change-Id: I4d41d44c0a83788e6e0f01bec3b59e60de8f12d4
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/8558
Reviewed-by: kristofer tomasette <kristofer_tomasette@comcast.com>
Tested-by: GerritAHP