Revamp loose Leader election
Previously, we stored only the ballot of the elected leader, which we
would use to instantiate an ActorRef whenever needed. Since that
instantiation is becoming more costly with the upcoming akka upgrade (or
at least its cost is becoming better known), we choose to save the Ref
as well as the ballot.

Also, switch from Option[Ballot] to more descriptive case
classes/objects for storing the current election state.

Change-Id: I843ff109607e9af395fcaa323e8ae7b0e67a74c0
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/8677
Tested-by: GerritAHP
Reviewed-by: kristofer tomasette <kristofer_tomasette@comcast.com>