Remove isNodeTerminated from SiriusInfo*
Unused field, the backing of which is deprecated in new versions. This
field is a bit of a leftover from a time when most of the questions
around "how to operationalize sirius" were still pretty open.

As it stands, the supervisor is only ever stopped when the impl itself
is shutdown, at which time the mbeanServer unregisters the info bean
(and the flag thus becomes inaccessible). Any other issue that would
"break" the supervisor would instead cause it to restart. I doubt this
flag has ever shown true in production.

Change-Id: Id8e1e2ac53fd978c04b56008c5ffab0c0bea40c5
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/9051
Tested-by: GerritAHP
Reviewed-by: Joe <joseph_campbell@comcast.com>