Replace .isTerminated with DeathWatch mechanic
.isTerminated is deprecated in newer versions of akka, replaced by the
mechanic of watching an actor and handling a Terminated message that
matches the actor's ref. It's actually a little cleaner, I think,
letting us keep more correct references to actors, and eliminating the
need for all the isTerminated calls.

But, y'know. We made a few of those calls here and there.

Change-Id: I4450a8079b411cfb491507a0b9ab490959bde704
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/9072
Reviewed-by: kristofer tomasette <kristofer_tomasette@comcast.com>
Tested-by: GerritAHP