Add repair() functionality to SegmentedUberstore
First, make SegmentedCompactor.replace() more atomic-op-friendly.
Deleting recursively is not atomic, and we could end up in a bad state
with no way to detect it (directory half-deleted) with replace()
implemented as before. This does the heavy lifting with move, which is
more of an atomic op.

Then, we're able to repair any possible state resulting from an
incomplete compaction or replace operation.

There's a little leakage from SegmentedCompactor into SegmentedUberStore
here (the method of doing the compact/replace, and the files that might
result), but I think we're probably ok with it. The alternative is to
move repair() into SegmentedCompactor, but I'm not sure I like that any
better.

Change-Id: Ie17bf04236cb7ccd511317eb0cb8309e50f91624
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/9432
Tested-by: GerritAHP
Reviewed-by: kristofer tomasette <kristofer_tomasette@comcast.com>