Add ClusterInfo for communicating cluster state
ClusterInfo is produced by the MembershipHelper (which may need a new
name). Right now, it's used by the Leader, when producing Scouts and
Commanders. As the datastructure held by the agent gets a little more
complex, it may make sense for most of the interactions with it (beyond
the MembershipActor) to be mediated through MembershipHelper.

So, getClusterInfo to MembershipHelper, wire the MembershipHelper in
lieu of the MembershipAgent into the Leader.

Change-Id: I97305f372e27f7829819311105aa964920670adb
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/9766
Tested-by: GerritAHP
Reviewed-by: kristofer tomasette <kristofer_tomasette@comcast.com>