Option up the ActorRef in MembershipAgent
Doesn't really have much of an effect currently, but with the upgrade,
any actor path that we find in the cluster config file will need to be
resolved with an Identify check into an actual ActorRef. When that's the
case, we need to be able to keep track of both the path and the current
state of the ActorRef resolution. This patchset expands the
MembershipAgent in this way.