Option up the ActorRef in MembershipAgent
Doesn't really have much of an effect currently, but with the upgrade,
any actor path that we find in the cluster config file will need to be
resolved with an Identify check into an actual ActorRef. When that's the
case, we need to be able to keep track of both the path and the current
state of the ActorRef resolution. This patchset expands the
MembershipAgent in this way.

Change-Id: I639f1cbdffc2cf03d156c3b0e9b9d4f777c14b4f
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/9771
Tested-by: GerritAHP
Reviewed-by: kristofer tomasette <kristofer_tomasette@comcast.com>