Add recursive chown of $WAL_DIR
We found that during the conversion process some directory
manipulation was done to save the old format of the wal dir.
This used the mv command, and the ownership becomes fncadm which
the tomcat process has no rights to files owned by fncadm. To fix
this a recursive chown was added (via sudo) to ensure the directory
of segmented wal files is owned by tomcat.

Change-Id: Ie772d875d0fe621ea4dc0b0eb7de523928276204
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/10218
Reviewed-by: kristofer tomasette <kristofer_tomasette@comcast.com>
Tested-by: kristofer tomasette <kristofer_tomasette@comcast.com>