Sirius should die in case of Config/WAL mismatch
Sirius will already fail to start if configured to instantiate a
SegmentedUberStore but handed a LegacyUberStore. However, in the
opposite case, it was not loudly failing -- it was instead building a
new Legacy UberStore in the root directory of the existing
SegmentedUberStore. This is a problem.

This, while not entirely palatable, causes a Legacy UberStore to check
for the Segmented VersionID, throwing an IllegalStateException if it's
found.