Sirius should die in case of Config/WAL mismatch
Sirius will already fail to start if configured to instantiate a
SegmentedUberStore but handed a LegacyUberStore. However, in the
opposite case, it was not loudly failing -- it was instead building a
new Legacy UberStore in the root directory of the existing
SegmentedUberStore. This is a problem.

This, while not entirely palatable, causes a Legacy UberStore to check
for the Segmented VersionID, throwing an IllegalStateException if it's
found.

Change-Id: I768c6d89fe5cfc48eed9e27f58cb50683673c9eb
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/10764
Tested-by: GerritAHP
Reviewed-by: kristofer tomasette <kristofer_tomasette@comcast.com>