Ignore duplicate Preempted messages
A Preempted message indicates that our ballot has been overridden by a
larger ballot, from a different Leader. We appropriately respond by
forwarding all of our active proposals over to the new leader. This is a
problem, however, if we receive multiple Preempted messages with the
same Ballot, which is possible (and even likely) if we have multiple
Proposals out and about in the system.

This change ignores Preempted messages if we have already processed and
identical one.