Ignore duplicate Preempted messages
A Preempted message indicates that our ballot has been overridden by a
larger ballot, from a different Leader. We appropriately respond by
forwarding all of our active proposals over to the new leader. This is a
problem, however, if we receive multiple Preempted messages with the
same Ballot, which is possible (and even likely) if we have multiple
Proposals out and about in the system.

This change ignores Preempted messages if we have already processed and
identical one.

Change-Id: I3a2ee3a44feb2f0bbdc36fa24fa4050f828c3ba5
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/11424
Tested-by: GerritAHP
Reviewed-by: Roger Chang <roger_chang@cable.comcast.com>
Reviewed-by: kristofer tomasette <kristofer_tomasette@comcast.com>