Introduce ClusterConfig trait and File-based impl
In retrospect, it was actually a little strange that the membership
actor was directly hitting the file on the disk. Additionally, it made
for some not-so-unity testing, since we had to read/write from the disk
while testing anything around membership.

This adds a thin layer of abstraction between the disk and the
MembershipActor.

The main goal for this change is that in upgrading akka versions, our
akka paths need to change slightly (akka => akka.tcp). To avoid breaking
anything, we'd like to introduce a temporary shim that can handle the
old-style akka paths, and that's just way too much low-level junk to ask
the MembershipActor to do. Plus, it also screwed with testing: we can't
replace akka:// with akka.tcp:// completely across the board; for
testing, when remoting is off, akka:// is still the correct prefix.

Change-Id: I4cae700887b3b0f6d2526e2fff42feb620db3f4c
Reviewed-on: https://source.dev.cim.comcast.net/gerrit/11553
Tested-by: GerritAHP
Reviewed-by: Roger Chang <roger_chang@cable.comcast.com>
Reviewed-by: kristofer tomasette <kristofer_tomasette@comcast.com>