Add support for akka over SSL.
This major change includes updating the sirius-akka-base-conf and all
the supporting classes.

Introduces new SiriusConfiguration elements including ENABLE_SSL,
SSL_RANDOM_NUMBER_GENERATOR, and elements dealing with
keystore/truststore locations and passwords.

SSL is off by default, and if turned on, uses java.security.SecureRandom
as the RNG by default.

This isn't battle-hardened. As it is, we've seen tests fail when using
AES RNGs on some hardware. Word on the street is that AES has a known
history of hanging on hardware where there's nothing else going on,
since it relies on FS changes to get its seed. We've not independently
confirmed this, however. In any case: Please bake thoroughly beforehand
if you're planning to use this in production.