add link requirement for pull requests
all pull requests should be linked to any issues they relate to.
Linking can happen in commit messages or the pull request message.
Not every commit message in a pull request needs a link to an issue,
but every pull request that is related to an issue should have a link.