Allow for a Sirius initialization callback.
Adds some messages to the SiriusSupervisor so that you can
send it a message to register interest in an initialization
event and get a message back once that has occurred. Added
an onInitialized() method to SiriusImpl to hook that up to
a callback.

Also defined a new trait/interface `Sirius1Dot2` that
extends `Sirius`. This seems like a good way to track
the additional methods without losing backwards
compatibility with version 1.1.4. In theory, this new
interface would be modifiable up until the next release
(which should be 1.2.0 in this case).

Discussion points:
* onInitialized takes a java.lang.Runnable as a callback to
  facilitate calling it from Java. Kinda kludgy, since from
  Scala it would just want a lazy `=> Unit`.
* Does the approach with the additional interface seem ok?