Make akka max-frame-size configurable
Also, make it actually do something. We currently specify it in a
non-functioning way for the version of akka we're using. Apparently the
key for that piece of configuration changed with the versions.

128K is the default for Akka, which is how it's been de facto configured
since the last release. Changing the default to 1024K, which is a little
more real-world useful for applications using Sirius, particularly WRT
catchup messages.

Also, make send/receive buffers scale with max-message-size.