Fix tty restoration logic in sirius-standalone
While this currently works the logic is off- we're checking to see
if saved_tty is empty, however saved_tty was never set. What we
meant to do was check if saved_stty is non-empty.

Additionally, move the exit portion out of the if as the if could be
not followed.