Increase FullSystemITest membership timeout
This has failed a couple of times, and we have a little evidence that
it's just barely missing. I don't like to just increase timeouts when we
see failures, but I'm not finding a better way to do it right now. Let's
increase the timeout for now, and if we still see failures, maybe take
another approach.