start a scout if remote leadership fails
when handling a leadership change, it is possible that sirius can fail
to resolve a remote reference to a new leader. This has been observed
during rolling restarts and on commodity hardware. If this happens and a
scout is not started, the node's leader will never resolve (ie remain
unknown) and will be unable to participate properly in its cluster.