Introduce .getDouble and .getInt to SiriusConfig
These attempt to read a value as the particular type, and if that fails,
to convert it to such. Not a ton of crazy logic, but since
SiriusConfiguration takes types (not just Strings), it's been kind of
like trying to hit a moving target sometimes to populate your config.
This makes it more flexible.