Move SiriusFactory to top-level api package.
This is a re-spin of #27 but also incorporates the intent of #68.
This should basically mean that a client of the Sirius library
ought only to directly use classes, objects, and traits from the
`com.comcast.xfinity.sirius.api` package, and that, going
forward, those are the files that must retain backwards
compatibility.

The main changes involved are:

* Additional extension methods for 1.3 are now collected in the
  Sirius1Dot3 trait that extends the existing Sirius trait.

* Existing SiriusFactory was hoisted up to api package and now
  returns a Sirius1Dot3 instead of a SiriusImpl. Clients ought
  to be depending on a trait rather than a concrete class here.

* There is still a SiriusFactory in the old location, but it
  just delegates to the one in the new location, then downcasts
  the return result to a SiriusImpl. This SiriusFactory is
  marked as deprecated.