Postpone StateBridge start until after Init
This effectively postpones catchup until after the WAL has been
replayed. Previously, we had considered benefits of doing these in
parallel, but for long replays (or heavy catchup needs), this can get a
little sketchy. Worse, it implicitly assumes that the environment can
handle it -- maybe not unreasonable, but unstated. This also opens the
way for some efficiencies in WAL replay (such as #49).

This has effects on "time to full startup" -- that is, time until replay
and catchup re both complete -- but I think it's worth it, for
simplicity and a simpler startup memory model.