Add liveness entry on resolution
MembershipActor tracks liveness (to some extent) of remote nodes by
relying on a Ping/Pong mechanism. It can evict dead-seeming nodes from
the MembershipMap when they haven't been "active" in a long time -- that
is, when it hasn't received a Pong for one of its Pings.

However, there's another time where we get information about ref
liveness, and that's on resolution. Currently there's an edge case where
a node can get a reference and not yet have any liveness data. Then, if
the ref immediately goes stale, there will be no way to evict it due to
unresponsiveness.

With this change, every "active" ref in the membership map should have
a corresponding last-liveness metric at all times.