Add backup_wal.sh and remove compacting dirs.
This change adds a new script, backup_wal.sh.  It is the equivalent of
compact_wal.sh, however it does not try to do any compacting.  It simply
makes a backup of uberstore.

The change also adds a step to both compact_wal.sh and backup_wal.sh to
remove any .compacting directories that exist.  These directories are
created during live compaction and are temporary.  If Sirius is stopped
during compaction, they may exist and if the timing is right they can
almost double the size of the uberstore.

They would be deleted when the first compaction starts anyway, so they
serve no useful purpose and can be deleted to save disk space.