Automated releases (#149)
To release, a person would go to the github releases tab, and create a new release. Travis ci will  then pick up the release as a tag and deploy using the `publish.sh` script.

The `publish.sh` script then uses the [sbt-sonatype](https://github.com/xerial/sbt-sonatype) plugin to publish, stage, and release to the public maven repository. The script overrides the version that is hard-coded inside `build.sbt` with the version in the tag during each release, so it is optional to have to update the version in `build.sbt` before releasing.

The account used for sonatype is Jonathan Ryan's. I'm using a user token so it's not my real credentials. I also created a GPG key to be used by travis for signing the release, that I've added to my github GPG keys. We may want to add other people's GPG keys, and sonatype credentials, and have each release's signature match the person who actually released it.